requirements.txt
torch
torchvision
nltk
Pillow
matplotlib
tensorflow==2.13.0
tensorflow-metal==0.6.0
